Move incremental fls
